add : AuthRequiredPage 컴포넌트 추가 및 스크랩, 마이페이지에 적용
